+++ vote/tests/test_front_end.py
@@ -1,3 +1,5 @@
+from .util import get_key_by_val, index_dict, modes_to_int, mode_aliases
+
 from django.test import LiveServerTestCase
 from selenium.webdriver.chrome.webdriver import WebDriver
 from selenium.webdriver.chrome.options import Options
@@ -96,43 +98,3 @@ class FrontEndTestCase(LiveServerTestCase):
         next_btn.click()
 
         sleep(1.5)
-
-
-def get_key_by_val(my_dict: dict, val: str or int):
-  for key, value in my_dict.items():
-    if val == value:
-      return key
-
-  raise Exception('Key doesn\'t exist')
-
-def index_dict(dictionary, n=0):
-  if n < 0:
-    n += len(dictionary)
-  for i, key in enumerate(dictionary.keys()):
-    if i == n:
-      return dictionary[key]
-  raise IndexError("dictionary index out of range")
-
-modes_to_int = {
-  'easy': 0,
-  'hard': 1,
-  'tematicas': 2,
-  'random_score': 3,
-  'min1': 4,
-  'min2': 5,
-  'deluxe': 6,
-  'end': 7,
-  'replica': 8,
-}
-
-mode_aliases = {
-  'easy': 'Easy Mode',
-  'hard': 'Hard Mode',
-  'tematicas': 'Temáticas',
-  'random_score': 'Random Mode',
-  'min1': 'Primer Minuto',
-  'min2': 'Segundo Minuto',
-  'deluxe': 'Deluxe',
-  'end': 'end',
-  'replica': 'Réplica',
-}
